Display selected product
The page displays a product when one of the tabs in the selection bar is
clicked. Since the code for each product is long I tried to create
separate files for each product and insert the corresponding file on the
main html using an click event listener to fetch the corresponding file
and inset where appropriate. However I realised that this is not so
simple as for example including an external tex file in a latex document.

The current solution requires all the html code to be in the same file,
I guess the above mentioned solution can be achieved, however I still
don't have the knowledge to build the solution.