Add logic to update shopping cart
When a new product is added to the shopping cart, a new objet of the
class product is created and added to the shopping cart. Additionally
a remove button and an input for the number of items of the same product
are shown. I faced the issue of reading an svg from a file, I found that
I cannot fetch files from my local machine and that a work around would
require running a local server. Instead I chose to add a dummy element
with display property set to hidden. Then I clone this element dynamically
to add an remove copies of it in the right place.