Add products to shopping cart display
To display the contents of the shopping cart I am using a dropdown
element to which I add a small picture of the item, the details: name
and flavor, the quantity, the price per item and the subtotal.

I tried to select the picture by targeting the element inside the
product displays. The selection worked, but the code showed an error
when I tried to clone the element. I still don't know what was the
issue. After googling for a fix without success, I decided to change the
approach and use a dictionary with the names of the image files to
insert. It seems that there is still some issue, because when I look
at the code flow on the dev tools, the flow stops when I load the image.
However the rest of the code continues and the product is displayed as
desired in the shopping cart. Gemini suggested that this could be due to
asynchronous image loading.