Enable buttons on shopping cart display
To modify the shopping cart using the buttons on the mini display,
we dispatch a click event on the corresponding buttons on the flavors
panel. To find the corresponding element and the object we want to
modify, I added the details of each product to the controls container
in both the flavors panel and the mini display.